[Browser Frame] Isolate Windows-only API
Windows browser frame needs to know where control buttons are so that
it can position frame elements and provide layout information to the
browser view.

However, this was shoved into DesktopWindowFrameHostWin, which involved
piping a single function specific to the Windows frame implementation
through a number of unrelated classes including BrowserWidget,
BrowserNativeWidget, BrowserDesktopWindowTreeHost implementations for
other platforms, etc.

This CL moves use of the helper class MinimizeButtonMetrics from
DesktopWindowTreeHostWin to BrowserFrameViewWin, removing a bunch of
otherwise unused plumbing and stub implementations. Because of the
way that MinimizeButtonMetrics works, a private impl class wraps it;
in a follow-up, we can unify those classes and perhaps move all of
the logic into the frame view private impl.

This follows the same pattern as all other platform browser frames,
which calculate/discover the location of caption buttons locally, as
they are frame elements.

Bug: 447104848
Change-Id: I2831ff76cd09abd53dc68788e4131b5898dd842c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6977465
Reviewed-by: Steven Luong <stluong@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520129}