[AIAE] Add the helper function to fetch APC from AEM
This CL added the helper functions to fetch the annotated page content
of a checkout page from AmountExtractionManager by calling
AutofillClient::GetAiPageContent() and save the data in the manager.
This data will be moved to the MES request once the
AmountExtractionManager is ready to make the server-side AI request.

To avoid multiple APC fetching requests racing with each other, if there
is a fetch ongoing or the data is already fetched, do not trigger the
new fetch.

This feature is protected by feature flag:
AutofillEnableAiBasedAmountExtraction and only available on Chrome
desktop for MVP.

This CL also added the amount_extraction_test_api.h to help testing. And
this CL also removed the platform condition checks, which will be added
back during launch.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Bug: 444685282
Change-Id: Iadffa5d7d176b62165d2bdc230614e1541e82a7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6969710
Reviewed-by: Olivia Saul <osaul@google.com>
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Commit-Queue: James Sun <jamessun@google.com>
Cr-Commit-Position: refs/heads/main@{#1520137}