Autofocus search box on returning to history page
When a user is in the history page's search mode, navigates to
another page, and then returns, the search box does not automatically
regain focus. This forces the user to tap the search box again to
continue typing.

This change addresses the issue by explicitly requesting focus on the
search EditText when the SelectableListLayout becomes visible.

- A new method `requestSearchFocus` is added to `SelectableListToolbar`
  to handle focusing the search input and showing the keyboard.
- `SelectableListLayout`'s `onWindowVisibilityChanged` method is
  overridden to call `requestSearchFocus` when the view becomes visible
  and is in search mode.

Bug: 422041670
Test: it should pass testSearchBarFocusOnSwitchToHistory unit test
Change-Id: I784fdc2302edefa743c443e13d5ebb885efff0bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6952914
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: ben chin <lungc@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520159}