[soft navs] Fix browser-initiated intercepted back navigations
The SoftNavigationHeuristics::EventScope created in
DocumentLoader::UpdateForSameDocumentNavigation() doesn't cover the
'navigate' event, which is used to intercept browser-initiated back
navs. The navigate event dispatch code (which is called into earlier)
does have logic to set up a new EventScope for user-initiated
navigations (which includes back navs), but the SoftNavigationHeruistics
code fails to create the scope because the event's isTrusted bit is
false. I think this used to work (> a year ago), but probably broke
with crrev.com/c/5458389.

This CL fixes the problem by unconditionally creating the EventScope
in that case, using CreateNavigationEventScope(), like we do in
DocumentLoader. This also adds:
 - Adds wpt_internal test covering this case. Unfortunately there isn't
   a test_driver API for this, so we'll have to make due with
   wpt_internal.
 - s/MaybeCreateEventScopeForEvent/MaybeCreateEventScopeForInputEvent,
   and removes the navigation event check, since this was broken.

Follow-up: With this change, there are actually two soft nav event
scopes: one created during the navigate event, and a nested one in
UpdateForSameDocumentNavigation. The SoftNavigationHeuristics logic
handles this case (reusing outer context), but the second one is
unnecessary, and slightly changes the start time. I'm not sure if this
was ever expected. The same can happen if there are nested navigate
events, e.g. calling pushState in a navigate event handler.

Fixed: 442852346
Change-Id: I0eae65077118b5d4c117e9782912442bdcc6b2d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6977508
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520163}