Call ThrottleManager directly from activation throttle
Until now, page activation was computed as follows:

(1) ThrottleManager adds an activation throttle. (2) The activation
throttle computes the page activation in WillProcessResponse and passes
it to a callback on the web contents helper. (3) The web contents helper
passes it to a callback on ThrottleManager.

This CL cuts out the unnecessary (2) step, so now it is just:

(1) ThrottleManager adds an activation throttle. (2) The activation
throttle computes the page activation in WillProcessResponse and passes
it to a callback on ThrottleManager.

Despite that the activation throttle doesn't actually throttle in
WillProcessResponse, it needs to be implemented as a NavigationThrottle
because the WillProcessResponse hook allows us to compute activation
only for navigation requests that successfully received a response.

Fixed: 346777548
Change-Id: Ib0c3b801459951e1e4300453fbaf2087d26a516b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6975520
Commit-Queue: Adam Abeshouse <aabeshouse@google.com>
Reviewed-by: Martin Pan-Verde <thesalsa@google.com>
Cr-Commit-Position: refs/heads/main@{#1520176}