[iOS] Add actions for Home Customization Background
This CL introduces user action metrics for the background customization
features.

The following actions are now being recorded:
- IOS.HomeCustomization.Background.PickerActionSheet.Tapped: When the
user taps the entry point to open the action sheet.
- IOS.HomeCustomization.Background.PickerActionSheet.Gallery.Tapped:
When the user selects gallery in the action sheet.
- IOS.HomeCustomization.Background.PickerActionSheet.UserUploaded.Tapped:
When the user selects upload image in the action sheet.
- IOS.HomeCustomization.Background.PickerActionSheet.Color.Tapped: When
the user selects color in the action sheet.
- IOS.HomeCustomization.Background.RecentlyUsed.Tapped: When the user
selects one of the recently used background images.
- IOS.HomeCustomization.Background.ResetDefault.Tapped: When the user
taps the option to reset the background to the default.

Bug: 444727712
Change-Id: I842b4306260d5bfd55c8a6139e71b417aafca7ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6968076
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Cheick Cisse <cheickcisse@google.com>
Cr-Commit-Position: refs/heads/main@{#1520201}