Move WebNN ORT to WinAppRuntime 1.8.1
WinML is part of WinAppRuntime 1.8.1 which shipped this week. Switch
Chromium to bind to the new package family name and update the
OpenVINO package family name.

A future change will source the EP package families from the
new-for-1.8.1 ExecutionProviderCatalog API in WinRT which avoids
the need to hardcode these.

Change-Id: I0edc7eaec9eb71d91eb1ec8f65d7907678ecd185
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6977675
Reviewed-by: ningxin hu <ningxin.hu@intel.com>
Commit-Queue: Rafael Cintron <rafael.cintron@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1520248}