media: Add Media.EME.{KeySystem}.LastKeyStatusMix to report the last key status
We report the initial key status only
(Media.EME.{KeySystem}.InitialKeyStatusMix) which is very helpful if
something is going on in the beginning. However, it's hard to debug
some issues where the key status is changed due to various reasons such
as opening/closing lid and connecting/disconnecting external monitors
during the playback.

There is a scenario where the initial is most likely all usable or all
pending but later we get another key status update callback with all or
partially output-restricted for some reasons. Since the user might end
up closing or refreshing the tab due to the error by output-restricted,
it would be helpful if we report the last key status UMA when the CDM
session gets closed.

Test results:
```
- Histogram: Media.EME.PlayReady.HardwareSecure.InitialKeyStatusMix recorded 3 samples, mean = 3.3 (flags = 0x41) [#]
0  -O                                                                        (1 = 33.3%)
1  ...
5  --O                                                                       (2 = 66.7%) {33.3%}
6  ...

- Histogram: Media.EME.PlayReady.HardwareSecure.LastKeyStatusMix recorded 3 samples, mean = 1.7 (flags = 0x41) [#]
0  --O                                                                       (2 = 66.7%)
1  ...
5  -O                                                                        (1 = 33.3%) {66.7%}
6  ...
```

Bug: 446727803
Change-Id: I1e7e0337ea6167cb9c55de21c3d8ff591f5d128f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6980426
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Sangbaek Park <sangbaekpark@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520274}