Add available memory threshold for spare renderer creation.
This change introduces a feature to prevent the creation of a spare
renderer on Android devices when the available memory is below a
certain threshold. The threshold is configurable via feature
parameters and is different for limited-memory and high-memory
devices.

This is part of the experimental android-only memory tracker for spare
renderer creation.

Bug: 443652225
Change-Id: Ibe31281b1e6904052cfd7248d754bc0d095f9a4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6978155
Commit-Queue: Jiacheng Guo <gjc@google.com>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520276}