[focusgroup] Add parsing support for opt out and memory.
This change adds the ability to parse the following tokens:

'none' => This will be used to allow focusable descendants of
focusgroups to "opt-out" from participating in the focusgroup.

`no-memory` => By default, focusgroup will restore the last-focused item
when re-entering. This option allows authors to disable this behavior.

This change does not implement the functionality, but simply parses
these tokens.

Bug: 40210717
Change-Id: Id89f4935197649a6fe74bc78ffbc8173c2658b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6969991
Commit-Queue: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520326}