[SxS] Ensure tab modal dialogs reposition correctly
When we are swapping a tab out of a split the series of events results
in the dialog being rendered over the left tab instead of the right.
This hook is one we should have been listening to in order to ensure we
position the dialog at the correct time.

Fixed: 445761437
Change-Id: I8edd03b78929f3dabb50d90b0f02ce506f020f7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6972844
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520328}