Fix: When minimized and restored, there is a flicker / glitch in the UI on macOS
When a minimized window exists and the user clicks the app’s Dock
icon, `applicationShouldHandleReopen` is triggered. Calling
`makeKeyAndOrderFront` on a minimized window causes it to be shown
immediately, then quickly hidden,before the proper restore animation
runs.  This results in a visual flash before the animation.

Change the logic to call `deminiaturize` when restoring minimized
windows, so that the window is brought back using the proper Dock
restore animation without the intermediate flash.

Fixed: 437807119
Change-Id: I211c08439b45d541e99785ce4ffd726734866922
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6975625
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Commit-Queue: hewro <ihewro@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1520347}