[SxS] Hide nudge when opening a new tab
When dropping a link from outside of chrome into the web contents, if
the split view drop target nudge was showing, after the link is dropped,
the nudge will still stay open. Fix this by hiding the nudge when we
observe that a tab was inserted into the tab strip. There are some cases
where the user can open a new tab while dragging (for example on Linux,
keyboard commands can be used during a drag), and in that case the nudge
will hide, but then it will just show again after the next drag update.

In the following screen recording, Canary (which has the bug) is on the
left, and my dev build with the fix is on the right:
https://screencast.googleplex.com/cast/NTk4NTg1MDc5NzA2NDE5MnwyMGJkMTRhYi1kMA

Fixed: 444712606
Change-Id: I3bd3226049f1fd8eed695e252ce8270e3adc0b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6974801
Reviewed-by: Alison Gale <agale@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520364}