Fall back to Receiver if DirectReceiver::Bind fails
Binding a DirectReceiver to another thread involves transferring the
PendingReceiver's endpoint to a ThreadLocalNode. This can theoretically
fail if the connection to that node closes, or more practically if the
async connection setup failed an error might be surfaced at the first
transfer attempt.

Binding the PendingReceiver to a Receiver in the global node doesn't
involve any extra connections so should succeed even if the
ThreadLocalNode transfer fails. The Receiver has the same functionality
as the DirectReceiver, just less efficient because it involves an extra
thread hop, so this patch treats DirectReceiver failures as non-fatal
and just uses a regular Receiver instead.

This also logs all failures to UMA so we can tell how often
DirectReceiver is falling back to Receiver.

NO_IFTTT=New IFTTT annotation.

Bug: 445243335
Change-Id: I0a4fc09f6d44b851301b500d7ded2f989892f24a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6974826
Commit-Queue: Joe Mason <joenotcharles@google.com>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Nafis Abedin <nafisabedin@google.com>
Cr-Commit-Position: refs/heads/main@{#1520378}