[Sync] Use kEncryptSyncCompat for OS crypt async encryption
This ensures that when the kSyncUseOsCryptAsync feature is enabled, the
data is encrypted with a scheme that is compatible with the older
encryption mechanism. This is important for safely rolling back the
feature flag, as it allows data encrypted with the new mechanism to be
decrypted by the old one.

Without this change, if the feature is enabled and then disabled, sync
data might become undecryptable, as the new encryption scheme (e.g.,
App-Bound on Windows) would be incompatible with the old decryption
mechanism.

Bug: 419157433
Change-Id: I7651ebbcc6c96ad429325a1187d28f72be5230fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6979649
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Michael Tatarski <mtatarski@google.com>
Commit-Queue: Michael Tatarski <mtatarski@google.com>
Cr-Commit-Position: refs/heads/main@{#1520476}