Avoid propagating GenerateStreams from inactive RFHs
Soft-fail GenerateStreams() calls from RFHs that are no longer active,
as is the case when navigating.

This is the short-term fix. In the long-term, it's also important to
associate the GenerateStreams message with the RFH in the rest of the
pipeline, as the RFH might still asynchronously deactivate at a later
time that is still before the dialog is shown to the user.

Bug: 442860743
Change-Id: Iaf322eb151ee6e916a24dd10468b2f0426216ac1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6966127
Commit-Queue: Elad Alon <eladalon@chromium.org>
Reviewed-by: Simon Hangl <simonha@google.com>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520479}