[reader mode] Add accessibility info to RM bottomsheet prefs buttons
Request from accessibility was to incorporate the font style and theme
buttons into MaterialButtonToggleGroup. However, integrating the buttons
as part of the MaterialButtonToggleGroup causes massive styling mismatch
from the desired design of the bottomsheet prefs. From testing locally
with the buttons in a ToggleGroup, the only significant difference was
the index of each button within the group (eg. 1 of 3). To prevent
having to override all the built-in ToggleGroup styling, this adds
AccessibilityDelegate to provide similar accessibility benefit.

Screen recording of buttons with updated accessibility:
http://screencast/cast/NTQ5OTYzNDc3OTY4NDg2NHxmYTBiMWFlNC03YQ

Bug: 444070753
Change-Id: I67fee5674b7c632cc1cd062866a997e039a613ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6981647
Commit-Queue: Evan Luo <evanluo@google.com>
Reviewed-by: Brandon Wylie <wylieb@google.com>
Cr-Commit-Position: refs/heads/main@{#1520527}