Make net::ProxyInfo take arguments by value.
And update a lot of callers to use move semantics.

Not expecting a huge win here, but on dev, we spend about 0.7
milliseconds in std::vector<net::ProxyChain>::push_back that() is not
used by the nested emplace_back() call, which I assume is spent
duplicating the ProxyChain, which is generally not needed.

Bug: 447146624
Change-Id: I0a56cb3b671e3f66b9aa33505cf7930eb5cca06b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6980824
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520538}