Include HiDPI resource pak files in webview builds.
This update was missed when we enabled hidpi resources on Android. The
change ensures higher quality static assets are available on WebView.

This change is effectively a no-op as it only permits the
`enable_hidpi` to be toggled to `true` on both Chrome and Webview
APKs.

Change-Id: I8569b6f53378728e42bb6f1b7f07c7073a45ebf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6853232
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1520588}