[remote_cocoa] Filter monitored events by window
This change introduces event filtering for monitored events in remote
cocoa windows.

The existing event monitoring code in NativeWidgetNSWindowBridge did
not filter events by the specific window, which caused some callers
to receive events for all windows.

To fix this, a `target_is_window` flag is now passed to
DispatchMonitorEvent to indicate if the event target matched the
correct window. This flag is bubbled up to the remote cocoa
EventMonitorMac implementation, where the correct filtering is now
done.

This fixes a previously skipped test in event_monitor_unittest.cc
and fixes a few regressions that happened when remote cocoa
EventMonitor support was originally implemented.

Bug: 426181639, 434135822
Change-Id: I0645c212a70afc8edc217556c9a883f5db74a220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6965320
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Fred Shih <ffred@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520637}