[omnibox][webui] Edit model notify webUI handler of selection
`OmniboxEditModel` will notify `WebuiOmniboxHandler`
`OnSelectionChanged().

Currently when `OmniboxEditModel` changes, it calls methods on its
`view_` and `popup_view_` directly to update their appearances
accordingly.

This works well with the assumption the edit model has a single `view_`
and `popup_view_` to update.

With the side-by-side debugging of the views and webUI omnibox popups,
`OmniboxEditModel` now has 2 popup views to update.

This CL introduces `OmniboxEditModelObserver`. Currently, it only
observes selection changes because that's the only change that the debug
webUI is interested in*. Other `OmniboxEditModel`

This affects both the debug and real `WebuiOmniboxHandler` since it'd be
weird to have one use the notification process while the other keeps
using a direct call through `popup_view_`. This does not the view
`OmniboxPopupViewViews`.

We can expand more of the direct calls from `OmniboxEditModel` to
`view_` and `popup_view_` to use the observer pattern in the future.
E.g. if we want the debug webUI popup to reflect drag updates, we would
have to replace `OmniboxPopupView::OnDragCanceled()` with a similar
observer method. But we can't completely replace the
`OmniboxEditModel::popup_view_` access because the `OmniboxEditModel`
also calls getters on the popup view.

*Actually, the debug webUI is obviously also interested in result
changes. But this is already achieved by observing the
`AutocompleteController`. Although `AutocompleteController` isn't owned
by the `OmniboxEditModel`, `AutocompleteController` does call
`OmniboxEditModel::OnPopupResultChanged()` when the
`AutocompleteResult`s change. So it would be nice to have the
`OmniboxEditModel` notify the debug webUI of both selection and result
changes; instead of the current state where the `OmniboxEditModel`
notifies for selection changes, while `AutocompleteController` notifies
for result changes. But this might be tricky since chrome://omnibox is
only interested in `AutocompleteResult` changes and doesn't care about
the larger `OmniboxEditModel` hierarchy.

Bug: 445425247
Change-Id: I48f33af767ce5a82749af6823cf54fe05e91d218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6981841
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Auto-Submit: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520645}