[Blink] Add check for context lost to CRPSI::RasterRecord()
It's not clear whether this is causing the issue in the linked bug, but
in any case it is an issue as the call to GetSyncToken() will crash if
the RasterInterface is not present.

This CL also defensively adds a check in CanvasResource::GetSyncToken(),
which we plan to remove later.

Bug: 447237526
Change-Id: I350da623003804191657327e68cc4cdeda85c553
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6983290
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520651}