[UNO] Update sync controls disabled/managed state
Until now, the toggles on the new account settings page were not
properly updated with sync disabled, as they were loaded before the
account settings page was shown and did not react to sync disabled. The
user had to refresh the pages in order for them to be properly disabled.

With this change, the toggles react to the signed in state rather than
the current route, which means that they are disabled with sync disabled
whenever the user is not syncing.

Additionally, the sync disabled information reacts to the current route.

This only affects changes behind the feature flag
`ReplaceSyncPromosWithSignInPromos`.

Bug: 446117071
Change-Id: I096ad0a69aee4054e34d3301881a2ce50c464be6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6973904
Reviewed-by: Ryan Sultanem <rsult@google.com>
Commit-Queue: Amelie Schneider <amelies@google.com>
Cr-Commit-Position: refs/heads/main@{#1520654}