[iOS] Add web view resizing type enum
This change add the `WebViewResizingType` enum to define different
resizing behaviors for the web view. This allows for more precise
control over how the viewport adapts to UI changes, such as the
appearance of toolbars or other elements.

This enum is a necessary component for the "Bling Fullscreen
refactoring" on iOS 26. It provides the infrastructure to handle various
fullscreen scenarios and ensures that content is resized correctly when
parts of the viewport are obscured.

See go/bling-fullscreen-refactoring-part1 for more details

This is part of multiples changes
* https://crrev.com/c/6839620 | [iOS] Implement fullscreen with obscured inset in iOS 26
* ➔ https://crrev.com/c/6853827 | [iOS] Add fullscreen resizing type enum (Current CL)
* https://crrev.com/c/6973698 | [iOS] Add web view resizing enum
* https://crrev.com/c/6827602 | [iOS] Add obscured inset property

Bug: 438144976
Change-Id: I310cc99ec1521bcd8929594a15d1e36187835911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6853827
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1520655}